requirements.txt
moviepy
speechrecognition
pydub
Pillow
whisper